+++ core-service/src/middleware/middleware.go
@@ -1,0 +1,11 @@
+package middleware
+
+// GoMiddleware represent the data-struct for middleware
+type GoMiddleware struct {
+	// another stuff , may be needed by middleware
+}
+
+// InitMiddleware initialize the middleware
+func InitMiddleware() *GoMiddleware {
+	return &GoMiddleware{}
+}
