+++ core-service/src/modules/event/repository/mysql/mysql_event.go
