+++ core-service/src/modules/information/repository/mysql/mysql_informations.go
