+++ core-service/src/modules/information/usecase/informations_ucase.go
