+++ core-service/src/modules/news/repository/mysql/mysql_news.go
