+++ core-service/src/modules/news/usecase/news_ucase.go
